+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## next
 
+- Exposed `createLoadDataState()` method as utils
 - Fixed displaying "Empty list" in view showcase when a view has no examples
 
 ## 1.0.0-beta.76 (10-07-2023)