+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed displaying "Empty list" in view showcase when a view has no examples
 - Fixed calling `onFinish()` callback on `Progressbar` getting an error
 - Fixed an exception on unsubscribing from progressbar state sync on loading data error
+- Fixed "uncaught error in promise" on data load failure in preloader
 - Removed deprecated option `title` in `expand` view config
 - Removed warning on definition `resolveLink()` method in page config